requirements.txt
arrow==1.2.3
glom==22.1.0
numpy==1.23.3
pandas==1.4.4
plotly==5.10.0
pycoingecko==3.0.0
PyYAML==6.0
requests==2.28.1
setuptools==65.1.1
SQLAlchemy==1.4.41
streamlit==1.12.2
openpyxl==3.0.10

python-dateutil~=2.8.2
python-dotenv~=0.21.0